+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# v7.4.1 (2024-05-21)
+
+Fixed
+- authCache needs namespace for each connection (#210)
+
+
+# v7.4.0 (2024-05-13)
+
+Added
+- Support for `INSERT OR IGNORE` (#207)
+- Support adding request timeout at query level (#208)
+
 # v7.3.0 (2024-04-17)
 
 Added